Feature/pay new endpoints

+++ src/main/java/com/talenthub/empresanominamicroservice/repository/ContractRepository.java
@@ -3,5 +3,5 @@ package com.talenthub.empresanominamicroservice.repository;
 import com.talenthub.empresanominamicroservice.model.Contract;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ContractRepository extends JpaRepository<Contract, Long> {
+public interface ContractRepository extends JpaRepository<Contract, Integer> {
 }
